+++ tests/scenario_tests/test_events_agent.py
@@ -1,10 +1,8 @@
 import json
 from time import sleep
-from unittest.mock import patch, MagicMock
 
 import pytest
 from slack_sdk.web import WebClient
-from slack_sdk.web.chat_stream import ChatStream
 
 from slack_bolt import App, BoltRequest, BoltContext, BoltAgent
 from slack_bolt.agent.agent import BoltAgent as BoltAgentDirect
@@ -57,90 +55,6 @@ class TestEventsAgent:
         assert response.status == 200
         assert_target_called()
 
-    def test_agent_chat_stream_uses_context_defaults(self):
-        """BoltAgent.chat_stream() passes context defaults to WebClient.chat_stream()."""
-        client = MagicMock(spec=WebClient)
-        client.chat_stream.return_value = MagicMock(spec=ChatStream)
-
-        agent = BoltAgentDirect(
-            client=client,
-            channel_id="C111",
-            thread_ts="1234567890.123456",
-            team_id="T111",
-            user_id="W222",
-        )
-        stream = agent.chat_stream()
-
-        client.chat_stream.assert_called_once_with(
-            channel="C111",
-            thread_ts="1234567890.123456",
-            recipient_team_id="T111",
-            recipient_user_id="W222",
-        )
-        assert stream is not None
-
-    def test_agent_chat_stream_overrides_context_defaults(self):
-        """Explicit kwargs to chat_stream() override context defaults."""
-        client = MagicMock(spec=WebClient)
-        client.chat_stream.return_value = MagicMock(spec=ChatStream)
-
-        agent = BoltAgentDirect(
-            client=client,
-            channel_id="C111",
-            thread_ts="1234567890.123456",
-            team_id="T111",
-            user_id="W222",
-        )
-        stream = agent.chat_stream(
-            channel="C999",
-            thread_ts="9999999999.999999",
-            recipient_team_id="T999",
-            recipient_user_id="U999",
-        )
-
-        client.chat_stream.assert_called_once_with(
-            channel="C999",
-            thread_ts="9999999999.999999",
-            recipient_team_id="T999",
-            recipient_user_id="U999",
-        )
-        assert stream is not None
-
-    def test_agent_chat_stream_rejects_partial_overrides(self):
-        """Passing only some of the four context args raises ValueError."""
-        client = MagicMock(spec=WebClient)
-        agent = BoltAgentDirect(
-            client=client,
-            channel_id="C111",
-            thread_ts="1234567890.123456",
-            team_id="T111",
-            user_id="W222",
-        )
-        with pytest.raises(ValueError, match="Either provide all of"):
-            agent.chat_stream(channel="C999")
-
-    def test_agent_chat_stream_passes_extra_kwargs(self):
-        """Extra kwargs are forwarded to WebClient.chat_stream()."""
-        client = MagicMock(spec=WebClient)
-        client.chat_stream.return_value = MagicMock(spec=ChatStream)
-
-        agent = BoltAgentDirect(
-            client=client,
-            channel_id="C111",
-            thread_ts="1234567890.123456",
-            team_id="T111",
-            user_id="W222",
-        )
-        agent.chat_stream(buffer_size=512)
-
-        client.chat_stream.assert_called_once_with(
-            channel="C111",
-            thread_ts="1234567890.123456",
-            recipient_team_id="T111",
-            recipient_user_id="W222",
-            buffer_size=512,
-        )
-
     def test_agent_available_in_action_listener(self):
         app = App(client=self.web_client)
 
@@ -192,16 +106,6 @@ class TestEventsAgent:
         response = app.dispatch(request)
         assert response.status == 200
         assert_target_called()
-
-    def test_agent_import_from_slack_bolt(self):
-        from slack_bolt import BoltAgent as ImportedBoltAgent
-
-        assert ImportedBoltAgent is BoltAgentDirect
-
-    def test_agent_import_from_agent_module(self):
-        from slack_bolt.agent import BoltAgent as ImportedBoltAgent
-
-        assert ImportedBoltAgent is BoltAgentDirect
 
     def test_agent_kwarg_emits_experimental_warning(self):
         app = App(client=self.web_client)
