+++ tests/slack_bolt/agent/test_agent.py
@@ -1,0 +1,103 @@
+from unittest.mock import MagicMock
+
+import pytest
+from slack_sdk.web import WebClient
+from slack_sdk.web.chat_stream import ChatStream
+
+from slack_bolt.agent.agent import BoltAgent
+
+
+class TestBoltAgent:
+    def test_chat_stream_uses_context_defaults(self):
+        """BoltAgent.chat_stream() passes context defaults to WebClient.chat_stream()."""
+        client = MagicMock(spec=WebClient)
+        client.chat_stream.return_value = MagicMock(spec=ChatStream)
+
+        agent = BoltAgent(
+            client=client,
+            channel_id="C111",
+            thread_ts="1234567890.123456",
+            team_id="T111",
+            user_id="W222",
+        )
+        stream = agent.chat_stream()
+
+        client.chat_stream.assert_called_once_with(
+            channel="C111",
+            thread_ts="1234567890.123456",
+            recipient_team_id="T111",
+            recipient_user_id="W222",
+        )
+        assert stream is not None
+
+    def test_chat_stream_overrides_context_defaults(self):
+        """Explicit kwargs to chat_stream() override context defaults."""
+        client = MagicMock(spec=WebClient)
+        client.chat_stream.return_value = MagicMock(spec=ChatStream)
+
+        agent = BoltAgent(
+            client=client,
+            channel_id="C111",
+            thread_ts="1234567890.123456",
+            team_id="T111",
+            user_id="W222",
+        )
+        stream = agent.chat_stream(
+            channel="C999",
+            thread_ts="9999999999.999999",
+            recipient_team_id="T999",
+            recipient_user_id="U999",
+        )
+
+        client.chat_stream.assert_called_once_with(
+            channel="C999",
+            thread_ts="9999999999.999999",
+            recipient_team_id="T999",
+            recipient_user_id="U999",
+        )
+        assert stream is not None
+
+    def test_chat_stream_rejects_partial_overrides(self):
+        """Passing only some of the four context args raises ValueError."""
+        client = MagicMock(spec=WebClient)
+        agent = BoltAgent(
+            client=client,
+            channel_id="C111",
+            thread_ts="1234567890.123456",
+            team_id="T111",
+            user_id="W222",
+        )
+        with pytest.raises(ValueError, match="Either provide all of"):
+            agent.chat_stream(channel="C999")
+
+    def test_chat_stream_passes_extra_kwargs(self):
+        """Extra kwargs are forwarded to WebClient.chat_stream()."""
+        client = MagicMock(spec=WebClient)
+        client.chat_stream.return_value = MagicMock(spec=ChatStream)
+
+        agent = BoltAgent(
+            client=client,
+            channel_id="C111",
+            thread_ts="1234567890.123456",
+            team_id="T111",
+            user_id="W222",
+        )
+        agent.chat_stream(buffer_size=512)
+
+        client.chat_stream.assert_called_once_with(
+            channel="C111",
+            thread_ts="1234567890.123456",
+            recipient_team_id="T111",
+            recipient_user_id="W222",
+            buffer_size=512,
+        )
+
+    def test_import_from_slack_bolt(self):
+        from slack_bolt import BoltAgent as ImportedBoltAgent
+
+        assert ImportedBoltAgent is BoltAgent
+
+    def test_import_from_agent_module(self):
+        from slack_bolt.agent import BoltAgent as ImportedBoltAgent
+
+        assert ImportedBoltAgent is BoltAgent
